requirements2.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
altair
tiktoken
huggingface-hub
InstructorEmbedding
sentence-transformers
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google-genai